+++ sliper/data/slips/sliper_app_data_processing_slips_main.py
@@ -1,60 +1,53 @@
 #!/usr/bin/python3
 """
-SLIPER APP - PREDICTORS PROCESSING - Soil Landslide Information and Prediction & Early Response
+SLIPER APP - SOIL SLIPS DATA PROCESSING - Soil Landslide Information and Prediction & Early Response
 
-__date__ = '20250724'
-__version__ = '1.4.0'
+__date__ = '20250714'
+__version__ = '1.0.0'
 __author__ =
-        'Stefania Magri (stefania.magri@arpal.liguria.it',
-        'Mauro Quagliati (mauro.quagliati@arpal.liguria.it',
-        'Monica Solimano (monica.solimano@arpal.liguria.it)',
-        'Fabio Delogu (fabio.delogu@cimafoundation.org'
+        'Fabio Delogu (fabio.delogu@cimafoundation.org',
+        'Francesco Silvestro (francesco.silvestro@cimafoundation.org)'
 
 __library__ = 'sliper'
 
 General command line:
-python app_predictors_main.py -settings_file configuration.json -time "YYYY-MM-DD HH:MM"
+python sliper_app_data_processing_slips_main.py -settings_file configuration.json -time "YYYY-MM-DD HH:MM"
 
 Version(s):
-20250724 (1.4.0) --> Refactor to the sliper package
-20221013 (1.3.0) --> Bugs fix to the kernel fx
-20220517 (1.2.0) --> Bugs fix and methods refactor
-20220320 (1.1.0) --> Beta release for operational chain mode
-20220105 (1.0.0) --> Beta release for Jupyter Notebook mode
+20250714 (1.0.0) --> Beta release for sliper package based on previous package(s) and version(s)
 """
 
-# -------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------
 # libraries
 import logging
 import time
 import os
 
 from driver_geo import DriverGeo
-from driver_training import DriverTraining
-from driver_predictors import DriverPredictors
+from driver_data import DriverData
 
 from argparse import ArgumentParser
 
-from lib_data_io_json import read_file_json
+from lib_data_io_json import read_file_settings
 from lib_utils_time import set_time
 from lib_utils_logging import set_logging_file
 from lib_info_args import logger_name, time_format_algorithm
 
-# Logging
+# logging
 log_stream = logging.getLogger(logger_name)
-# -------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------
 
-# -------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------
 # Algorithm information
 project_name = 'sliper'
-alg_name = 'SLIPER APP - PREDICTORS PROCESSING'
+alg_name = 'SLIPER APP - SLIPS DATA PROCESSING'
 alg_type = 'Package'
-alg_version = '1.4.0'
-alg_release = '2025-07-24'
-# -------------------------------------------------------------------------------------
+alg_version = '1.0.0'
+alg_release = '2025-07-14'
+# ----------------------------------------------------------------------------------------------------------------------
 
 
-# -------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------
 # Script Main
 def main():
 
@@ -63,7 +56,7 @@ def main():
     alg_settings, alg_time = get_args()
 
     # Set algorithm settings
-    data_settings = read_file_json(alg_settings)
+    data_settings = read_file_settings(alg_settings)
 
     # Set algorithm logging
     set_logging_file(
@@ -93,55 +86,42 @@ def main():
         time_format=time_format_algorithm,
         time_period=data_settings['time']['time_period'],
         time_frequency=data_settings['time']['time_frequency'],
-        time_rounding=data_settings['time']['time_rounding'],
-        time_reverse=False
+        time_rounding=data_settings['time']['time_rounding']
     )
     # ------------------------------------------------------------------------------------------------------------------
 
     # ------------------------------------------------------------------------------------------------------------------
     # Geographical datasets
     driver_geo = DriverGeo(
-        src_dict=data_settings['data']['static']['source']['geo_datasets'],
-        dst_dict=data_settings['data']['static']['destination']['geo_datasets'],
+        src_dict=data_settings['data']['static']['source'],
+        dst_dict=data_settings['data']['static']['destination'],
         tmp_dict=data_settings['tmp'],
         tags_dict=data_settings['algorithm']['template'],
         flag_update=data_settings['algorithm']['flags']['update_static'])
     # organize geo collections
-    geo_data = driver_geo.organize_data()
+    geo_info, geo_data, geo_areas = driver_geo.organize_data()
     # ------------------------------------------------------------------------------------------------------------------
 
-    # -------------------------------------------------------------------------------------
-    # Training datasets
-    driver_training = DriverTraining(
-        src_dict=data_settings['data']['static']['source']['training_datasets'],
-        dst_dict=data_settings['data']['static']['destination']['training_datasets'],
-        flag_update=data_settings['algorithm']['flags']['update_static'])
-    training_data = driver_training.organize_data()
-    # -------------------------------------------------------------------------------------
-
-    # -------------------------------------------------------------------------------------
-    # driver to define indicators
-    driver_predictors = DriverPredictors(
-        time_run, time_range,
+    # ------------------------------------------------------------------------------------------------------------------
+    # method to call driver data
+    driver_data = DriverData(
+        time_run=time_run,
         src_dict=data_settings['data']['dynamic']['source'],
         anc_dict=data_settings['data']['dynamic']['ancillary'],
         dst_dict=data_settings['data']['dynamic']['destination'],
         tags_dict=data_settings['algorithm']['template'],
-        parameters_dict=data_settings['algorithm']['parameters'],
-        geo_data=geo_data,
-        training_data=training_data,
-        flag_update_anc=data_settings['algorithm']['flags']['update_dynamic_ancillary'],
-        flag_update_dst=data_settings['algorithm']['flags']['update_dynamic_destination']
-    )
-    # method to organize data collections
-    data_collections = driver_predictors.organize_data()
-    # method to analyze data collections
-    analysis_datasets = driver_predictors.compute_data(data_collections)
-    # method to dump data collections
-    driver_predictors.dump_data(analysis_datasets)
-    # -------------------------------------------------------------------------------------
+        tmp_dict=data_settings['tmp'],
+        collections_data_geo_info=geo_info,
+        collections_data_geo_ref=geo_data,
+        collections_data_geo_areas=geo_areas,
+        flag_update=data_settings['algorithm']['flags']['update_dynamic'])
 
-    # -------------------------------------------------------------------------------------
+    # organize data collections
+    data_info = driver_data.organize_data()
+    # dump data collections
+    driver_data.dump_data(data_info)
+
+    # --------------------------------------------------------------------------------------------------------------
 
     # ------------------------------------------------------------------------------------------------------------------
     # Info algorithm
@@ -158,8 +138,9 @@ def main():
 
 # ----------------------------------------------------------------------------------------------------------------------
 
+
 # ----------------------------------------------------------------------------------------------------------------------
-# Method to get script argument(s)
+# method to get script argument(s)
 def get_args():
     parser_handle = ArgumentParser()
     parser_handle.add_argument('-settings_file', action="store", dest="alg_settings")
@@ -176,9 +157,9 @@ def get_args():
 
 # ----------------------------------------------------------------------------------------------------------------------
 
+
 # ----------------------------------------------------------------------------------------------------------------------
 # Call script from external library
 if __name__ == '__main__':
     main()
 # ----------------------------------------------------------------------------------------------------------------------
-