+++ sliper/data/sm/sliper_app_data_processing_sm_main.py
@@ -11,7 +11,7 @@ __author__ =
 __library__ = 'sliper'
 
 General command line:
-python app_data_processing_sm_main.py -settings_file configuration.json -time "YYYY-MM-DD HH:MM"
+python sliper_app_data_processing_sm_main.py -settings_file configuration.json -time "YYYY-MM-DD HH:MM"
 
 Version(s):
 20250618 (1.0.0) --> Beta release for sliper package based on previous package(s) and version(s)